+++ src/api/index.ts
@@ -1,3 +1,5 @@
-export * from './loadSupport'
-export * from './runCucumber'
+export * from './convert_configuration'
+export * from './load_configuration'
+export * from './load_support'
+export * from './run_cucumber'
 export * from './types'
