+++ src/cli/helpers.ts
@@ -238,3 +238,10 @@ export function emitSupportCodeMessages({
   emitTestCaseHooks(supportCodeLibrary, eventBroadcaster)
   emitTestRunHooks(supportCodeLibrary, eventBroadcaster)
 }
+
+export function isTruthyString(s: string | undefined): boolean {
+  if (s === undefined) {
+    return false
+  }
+  return s.match(/^(false|no|0)$/i) === null
+}
